[ska][x-pack] Move painless_lab, upgrade_assistant and reporting_management functional tests (#225829)
## Summary

Part of https://github.com/elastic/kibana-team/issues/1503

This PR relocates `painless_lab `, `upgrade_assistant ` and
`reporting_management ` test files from `x-pack/test/functional/apps`
dir to `x-pack/platform/test/functional/apps`.

Before:

```
/x-pack/test/functional/apps
     | - painless_lab/
     | - upgrade_assistant/
     | - reporting_management/
```

After:
```
/x-pack/platform/test/functional/apps
     | - painless_lab/
     | - upgrade_assistant/
     | - reporting_management/
```

---------

(cherry picked from commit 8ccd09fef97e0bd51e85ae2c8adb9166d37eaef1)
