+++ x-pack/platform/test/functional/apps/painless_lab/painless_lab_flyout.ts
