+++ x-pack/platform/test/functional/apps/reporting_management/index.ts
@@ -4,8 +4,9 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+import { FtrProviderContext } from '../../ftr_provider_context';
 
-export default ({ loadTestFile }) => {
+export default ({ loadTestFile }: FtrProviderContext) => {
   describe('reporting management app', function () {
     loadTestFile(require.resolve('./report_listing'));
   });
