+++ x-pack/platform/test/functional/page_objects/index.ts
@@ -18,6 +18,7 @@ import { GisPageObject } from './gis_page';
 import { ReportingPageObject } from './reporting_page';
 import { WatcherPageObject } from './watcher_page';
 import { SearchProfilerPageProvider } from './search_profiler_page';
+import { UpgradeAssistantFlyoutObject } from './upgrade_assistant_page';
 
 // just like services, PageObjects are defined as a map of
 // names to Providers. Merge in Kibana's or pick specific ones
@@ -35,4 +36,5 @@ export const pageObjects = {
   reporting: ReportingPageObject,
   watcher: WatcherPageObject,
   searchProfiler: SearchProfilerPageProvider,
+  upgradeAssistant: UpgradeAssistantFlyoutObject,
 };
